ORMs/Tortoise_ORM/test_lab_6/requirements.txt
aiosqlite==0.17.0
anyio==3.7.1
pypika-tortoise==0.1.6
pytest==7.4.0
pytest-asyncio==0.21.0
python-dateutil==2.8.2
pytz==2023.3
tortoise-orm==0.19.3
typing_extensions==4.7.1
tzdata==2023.3
